% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markers.R
\name{build_dotplot}
\alias{build_dotplot}
\title{Builds a data frame of all genes and their markers to be plotted}
\usage{
build_dotplot(all_markers, top_markers, mat_norm, cell_type)
}
\arguments{
\item{all_markers}{A DataFrame of all cells including genes, and cell-type.}

\item{top_markers}{A DataFrame of the top markers per gene}

\item{mat_norm}{Large dgeMatrix of the normalized counts of the SingleCellExperiment object}

\item{cell_type}{A dataframe of the filtered genes and all of their respective cell types}
}
\value{
A dataframe of each filtered gene, ordered by cell type.
}
\description{
Builds a data frame of all genes and their markers to be plotted
}
