% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Example SingleCellExperiment dataset}
\format{
A \code{SingleCellExperiment} object.
}
\source{
\code{scRNAseq::BaronPancreasData("mouse")}
}
\usage{
example_sce
}
\description{
A filtered example dataset based on the Baron mouse pancreas
single-cell RNA-seq data.
}
\keyword{datasets}
