% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_markers.R
\name{export_markers}
\alias{export_markers}
\title{Exports all markers}
\usage{
export_markers(all_markers, marker_summary, output_dir)
}
\arguments{
\item{all_markers}{A DataFrame of all cells including genes, and cell-type.}

\item{marker_summary}{A dataframe with all distinct cell types with their total markers and top marker.}

\item{output_dir}{Creates a directory for the data to be exported in}
}
\value{
An exported table of all the genes and their markers.
}
\description{
Exports all markers
}
