% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_analysis.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes by detection rate}
\usage{
filter_genes(sce, min_detect_rate = 0.05)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{min_detect_rate}{Minimum fraction of cells in which a gene must be detected.}
}
\value{
A filtered SingleCellExperiment object.
}
\description{
Filter genes by detection rate
}
\examples{
\dontrun{
data(example_sce)
sce_filtered <- filter_genes(example_sce, min_detect_rate = 0.05)
}
}
