% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_analysis.R
\name{find_markers}
\alias{find_markers}
\title{Filter genes by detection rate}
\usage{
find_markers(sce, mat_norm, group_col = "label")
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{min_detect_rate}{Minimum fraction of cells in which a gene must be detected.}
}
\value{
A filtered SingleCellExperiment object.
}
\description{
Filter genes by detection rate
}
