% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cli.R
\name{install_project12pkg_cli}
\alias{install_project12pkg_cli}
\title{Install CLI launchers for project12pkg}
\usage{
install_project12pkg_cli()
}
\value{
Invisibly returns the result of \code{Rapp::install_pkg_cli_apps()}.
}
\description{
Installs command-line launchers for the package CLI app.
}
