% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_analysis.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize gene counts}
\usage{
normalize_counts(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}
}
\value{
A normalized SingleCellExperiment object.
}
\description{
Normalize gene counts
}
