% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_analysis.R
\name{select_top_markers}
\alias{select_top_markers}
\title{Select top N markers per cell type}
\usage{
select_top_markers(all_markers, n_markers = 5)
}
\arguments{
\item{all_markers}{A DataFrame of all cells including genes, and cell-type.}

\item{n_markers}{Preferred number of markers.}
}
\value{
A cell_types object with inputted number of top markers per cell types.
}
\description{
Select top N markers per cell type
}
\examples{
\dontrun{
data(example_sce)
mat_norm <- normalize_counts(example_sce)
all_markers <- find_markers(example_sce, mat_norm, group_col = "label")
top_markers <- select_top_markers(all_markers, n_markers = 5)
}
}
