% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_markers.R
\name{summarize_markers}
\alias{summarize_markers}
\title{Summary: overlap counts between groups}
\usage{
summarize_markers(all_markers, padj_cutoff = 1, log2fc_cutoff = 1)
}
\arguments{
\item{all_markers}{A DataFrame of all cells including genes, and cell-type.}

\item{padj_cutoff}{A value for the desired cut off for the p-adjusted value}

\item{log2fc_cutoff}{A value for the desired cut off for the log2fc value.}
}
\value{
A dataframe with all distinct cell types with their total markers and top marker.
}
\description{
Summary: overlap counts between groups
}
\examples{
\dontrun{
data(example_se)

# Example
# marker_summary <- summarize_markers(...)
}
}
