% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_analysis.R
\name{find_markers}
\alias{find_markers}
\title{Filter genes by detection rate}
\usage{
find_markers(sce, mat_norm, group_col = "label")
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{mat_norm}{A normalized expression matrix.}

\item{group_col}{Column name in \code{colData(sce)} containing cell type labels.}
}
\value{
A filtered SingleCellExperiment object.
}
\description{
Filter genes by detection rate
}
