% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_analysis.R
\name{find_markers}
\alias{find_markers}
\title{Find Marker genes for each cell type}
\usage{
find_markers(sce, mat_norm, group_col = "label")
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{mat_norm}{A normalized expression matrix.}

\item{group_col}{Column name in \code{colData(sce)} containing cell type labels.}
}
\value{
A data.frame of marker statistics for each gene and cell type.
}
\description{
Find Marker genes for each cell type
}
\examples{
\dontrun{
data(example_sce)
mat_norm <- normalize_counts(example_sce)
all_markers <- find_markers(example_sce, mat_norm, group_col = "label")
}
}
