% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markers.R
\name{plot_markers}
\alias{plot_markers}
\title{Plots genes and their markers}
\usage{
plot_markers(plot_df)
}
\arguments{
\item{plot_df}{A dataframe of all filtered genes}
}
\value{
A ggplot dotplot of the plot_df data frame
}
\description{
Plots genes and their markers
}
\examples{
\dontrun{
data(example_se)

# Example workflow
# plot_df <- build_dotplot(...)

# Plot markers
plot_markers(plot_df)
}
}
